chore: move "cli-kinton delete" to "cli-kintone record delete"
